image-processing/requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.0
idna==3.10
imageio==2.37.0
jinja2==3.1.6
lazy-loader==0.4
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
sympy==1.14.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.3
typing==3.10.0.0
typing-extensions==4.14.1
urllib3==2.5.0
